unsigned/uint32_t/uint64_t cleanup

+++ src/btoraig.h
@@ -61,15 +61,15 @@ struct BtorAIGMgr
   BtorAIGPtrStack id2aig; /* id to AIG node */
   BtorIntStack cnfid2aig; /* cnf id to AIG id */
 
-  long long cur_num_aigs;     /* current number of ANDs */
-  long long cur_num_aig_vars; /* current number of AIG variables */
+  uint_least64_t cur_num_aigs;     /* current number of ANDs */
+  uint_least64_t cur_num_aig_vars; /* current number of AIG variables */
 
   /* statistics */
-  long long max_num_aigs;
-  long long max_num_aig_vars;
-  long long num_cnf_vars;
-  long long num_cnf_clauses;
-  long long num_cnf_literals;
+  uint_least64_t max_num_aigs;
+  uint_least64_t max_num_aig_vars;
+  uint_least64_t num_cnf_vars;
+  uint_least64_t num_cnf_clauses;
+  uint_least64_t num_cnf_literals;
 };
 
 typedef struct BtorAIGMgr BtorAIGMgr;
